pyproject.toml
[project]
name = "course-coosing-jw-ustc"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "铂屑", email = "80164656+boxie123@users.noreply.github.com" }
]
dependencies = [
    "httpx>=0.28.1",
    "prettytable>=3.12.0",
    "jmespath>=1.0.1",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/course_coosing_jw_ustc"]
